refactor: add getConfigFile function
